[ALPS09092314] disp: fix coverity issue
[Description]
1. check null pointer before use
2. force size before dump
3. check index before access array

[Test]
display MEstress

MTK-Commit-Id: a90070832cf5be50fcee7d30a9bbf9f6f0ba1826

CR-Id: ALPS09092314
Feature: display_mobile

+++ drivers/gpu/drm/mediatek/mediatek_v2/mtk_drm_lowpower.c
@@ -379,7 +379,7 @@ void mtk_drm_idlemgr_cpu_control(struct drm_crtc *crtc, int cmd, unsigned int da
 
 	DDPMSG("%s,crtc:%u mask:0x%x, freq:%uMhz, latency:%dus\n", __func__,
 		crtc_id, idlemgr_ctx->priv.cpu_mask,
-		idlemgr_ctx->priv.cpu_freq / 1000,
+		(unsigned int)(idlemgr_ctx->priv.cpu_freq / 1000),
 		idlemgr_ctx->priv.cpu_dma_latency);
 	DDP_MUTEX_UNLOCK(&mtk_crtc->lock, __func__, __LINE__);
 }
@@ -1806,7 +1806,8 @@ int mtk_drm_idlemgr_init(struct drm_crtc *crtc, int index)
 		DDPMSG("%s, %d, crtc:%u,cmd:%d,async:%d/%d,cpu:0x%x-%uMhz,sram:%d\n",
 			__func__, __LINE__, drm_crtc_index(crtc), mode,
 			idlemgr_ctx->priv.hw_async, idlemgr_ctx->priv.vblank_async,
-			idlemgr_ctx->priv.cpu_mask, idlemgr_ctx->priv.cpu_freq / 1000,
+			idlemgr_ctx->priv.cpu_mask,
+			(unsigned int)(idlemgr_ctx->priv.cpu_freq / 1000),
 			idlemgr_ctx->priv.sram_sleep);
 		mtk_drm_idlemgr_bind_cpu(idlemgr->idlemgr_task, crtc, true);
 		mtk_drm_idlemgr_bind_cpu(idlemgr->kick_task, crtc, true);
@@ -2076,7 +2077,7 @@ static void mtk_drm_idlemgr_disable_crtc(struct drm_crtc *crtc)
 					"%s:async:%d,cpu:(0x%x,%uMhz,%dus),sram:%d,total:%lluus,detail:%s\n",
 					__func__, atomic_read(&idlemgr->async_enabled),
 					idlemgr_ctx->priv.cpu_mask,
-					idlemgr_ctx->priv.cpu_freq / 1000,
+					(unsigned int)(idlemgr_ctx->priv.cpu_freq / 1000),
 					idlemgr_ctx->priv.cpu_dma_latency,
 					idlemgr_ctx->priv.sram_sleep,
 					cost, perf_string);
@@ -2349,7 +2350,7 @@ static void mtk_drm_idlemgr_enable_crtc(struct drm_crtc *crtc)
 					"%s:async:%d,cpu:(0x%x,%uMhz,%dus),sram:%d,total:%lluus,detail:%s\n",
 					__func__, atomic_read(&idlemgr->async_enabled),
 					idlemgr_ctx->priv.cpu_mask,
-					idlemgr_ctx->priv.cpu_freq / 1000,
+					(unsigned int)(idlemgr_ctx->priv.cpu_freq / 1000),
 					idlemgr_ctx->priv.cpu_dma_latency,
 					idlemgr_ctx->priv.sram_sleep,
 					cost, perf_string);
@@ -2358,7 +2359,7 @@ static void mtk_drm_idlemgr_enable_crtc(struct drm_crtc *crtc)
 
 		if (perf_aee_timeout > 0 && cost > (unsigned long long)perf_aee_timeout * 1000) {
 			DDPAEE("[IDLE] perf drop:%lluus, timeout:%uus\n",
-				cost, perf_aee_timeout * 1000);
+				cost, (unsigned int)(perf_aee_timeout * 1000));
 			perf_aee_timeout = 0;
 		}
 	}
